Revert "fix: move tests closer to implementation"
This reverts commit b269c133be1d2a649fe29ad6ca1cf32fc98def7b.
