+++ src/core/merge_right.rs
@@ -477,69 +477,4 @@ mod tests {
 
         assert_eq!(a.merge_right(b), expected);
     }
-
-    mod eval {
-        use async_graphql_value::ConstValue;
-        use serde_json::json;
-
-        use crate::core::blueprint::{Blueprint, DynamicValue};
-        use crate::core::http::RequestContext;
-        use crate::core::ir::model::IR;
-        use crate::core::ir::{EmptyResolverContext, EvalContext};
-
-        #[tokio::test]
-        async fn test_const_values() {
-            let a = DynamicValue::Value(
-                ConstValue::from_json(json!({
-                    "a": 1,
-                    "c": {
-                        "ca": false
-                    }
-                }))
-                .unwrap(),
-            );
-
-            let b = DynamicValue::Value(
-                ConstValue::from_json(json!({
-                    "b": 2,
-                    "c": {
-                        "cb": 23
-                    }
-                }))
-                .unwrap(),
-            );
-
-            let c = DynamicValue::Value(
-                ConstValue::from_json(json!({
-                    "c" : {
-                        "ca": true,
-                        "cc": [1, 2]
-                    },
-                    "d": "additional"
-                }))
-                .unwrap(),
-            );
-
-            let ir = IR::Merge([a, b, c].into_iter().map(IR::Dynamic).collect());
-            let runtime = crate::cli::runtime::init(&Blueprint::default());
-            let req_ctx = RequestContext::new(runtime);
-            let res_ctx = EmptyResolverContext {};
-            let mut eval_ctx = EvalContext::new(&req_ctx, &res_ctx);
-
-            let actual = ir.eval(&mut eval_ctx).await.unwrap();
-            let expected = ConstValue::from_json(json!({
-                "a": 1,
-                "b": 2,
-                "c": {
-                    "ca": true,
-                    "cb": 23,
-                    "cc": [1, 2]
-                },
-                "d": "additional"
-            }))
-            .unwrap();
-
-            assert_eq!(actual, expected);
-        }
-    }
 }
